dashboard

+++ lib/data/constants/emotions.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 
 enum Emotion {
-  happy(emoji: '😊', label: 'Happy', color: Colors.yellow, value: 5),
-  sad(emoji: '😢', label: 'Sad', color: Colors.blue, value: 4),
-  calm(emoji: '😌', label: 'Calm', color: Colors.green, value: 3),
-  fear(emoji: '😨', label: 'Fear', color: Colors.orange, value: 2),
-  angry(emoji: '😠', label: 'Angry', color: Colors.red, value: 1);
+  happy(emoji: '😊', label: 'Happy', color: Colors.yellow, value: 5.0),
+  sad(emoji: '😢', label: 'Sad', color: Colors.blue, value: 4.0),
+  calm(emoji: '😌', label: 'Calm', color: Colors.green, value: 3.0),
+  fear(emoji: '😨', label: 'Fear', color: Colors.orange, value: 2.0),
+  angry(emoji: '😠', label: 'Angry', color: Colors.red, value: 1.0);
 
   const Emotion({
     required this.emoji,
@@ -17,5 +17,5 @@ enum Emotion {
   final String emoji;
   final String label;
   final Color color;
-  final int value;
+  final double value;
 }
